requirements.txt
torch==2.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
huggingface-hub==0.19.4
safetensors==0.4.1
Pillow==10.1.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.0
numpy==1.24.4
omegaconf==2.3.0
aiohttp==3.9.1
discord.py==2.3.2
google-genai
xformers==0.0.22
compel